requirements.txt
aiohttp==3.9.3
discord.py==2.4.0
